+++ Hubanghu/Hubanghu/TwoViewController/Model/HbhWorkers.h
@@ -12,10 +12,10 @@
 @interface HbhWorkers : NSObject <NSCoding, NSCopying>
 
 @property (nonatomic, strong) NSString *name;
-@property (nonatomic, assign) double workersIdentifier;
+@property (nonatomic, assign) int workersIdentifier;
 @property (nonatomic, strong) NSString *workTypeName;
 @property (nonatomic, strong) NSString *photoUrl;
-@property (nonatomic, assign) double orderCount;
+@property (nonatomic, assign) int orderCount;
 @property (nonatomic, strong) NSArray *caseProperty;
 @property (nonatomic, strong) NSString *workingAge;
 @property (nonatomic, strong) NSArray *comment;
